requirements-test.txt
# Test Dependencies

# Core testing frameworks
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=3.0.0
pytest-watch>=4.2.0
pytest-timeout>=2.1.0

# Mocking and fixtures
factory-boy>=3.2.0
faker>=13.0.0
mixer>=4.0.0

# Django and DRF testing
django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=3.14.0

# Additional utilities
python-dotenv>=0.20.0
coverage>=6.0

# Performance and quality
pytest-benchmark>=3.4.0
pytest-xdist>=2.5.0  # For parallel test execution
